pyproject.toml
[tool.poetry]
name = "fundamentals-of-nlp"
version = "0.1.0"
description = ""
authors = ["masayukeeeee <masayukeeeee@gmail.com>"]
readme = "README.md"
packages = [
  { include = "lib"}
]

[tool.poetry.dependencies]
python = "^3.11"
flake8 = "^7.0.0"
pytest = "^7.4.4"
mypy = "^1.8.0"
ipdb = "^0.13.13"
mecab-python3 = "^1.0.8"
pandas = "^2.2.0"
pyarrow = "^15.0.0"
polars = "^0.20.6"

[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.0.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


